Python_ORM

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,40 +1,33 @@
-#!/usr/bin/python
-"""a script that deletes all State objects with a name containing the 
-letter a from the database hbtn_0e_6_usa
+#!/usr/bin/python3
 """
-import sys
+This script deletes all State objects
+with a name containing the letter `a`
+from the database `hbtn_0e_6_usa`.
+"""
+
+from sys import argv
+from model_state import State, Base
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from model_state import Base, State
-
 
 if __name__ == "__main__":
-    # Get command-line arguments
-    username = sys.argv[1]
-    password = sys.argv[2]
-    database = sys.argv[3]
+    """
+    Deletes State objects on the database.
+    """
 
-    # Create the engine to connect to the MySQL server
-    engine = create_engine
-    (f"mysql://{username}:{password}@localhost:3306/{database}")
+    db_url = "mysql+mysqldb://{}:{}@localhost:3306/{}".format(
+        argv[1], argv[2], argv[3])
 
-    # Bind the engine to the Base class
-    Base.metadata.bind = engine
-
-    # Create a session to interact with the database
+    engine = create_engine(db_url)
     Session = sessionmaker(bind=engine)
+
     session = Session()
 
-    # Query the State objects with names containing the letter "a"
-    states_to_delete = session.query(State).filter
-    (State.name.like("%a%")).all()
+    states = session.query(State).filter(State.name.contains('a'))
+    if states is not None:
+        for state in states:
+            session.delete(state)
 
-    # Delete the matching states
-    for state in states_to_delete:
-        session.delete(state)
-
-    # Commit the changes
     session.commit()
 
-    # Close the session
     session.close()
